The moving mesh track nearly working. Occasionally a hit is not detected and the particle moves a long way outside the tet. A better system of rejecting roots is needed, probably based on the differential of the hit coordinate. This might also be applicable to the stationary mesh case.
Tracking through moving degenerate tets has not been tested, but the
moving analysis is done in terms of mu, rather than lambda, so it should
work.

Cases where the tet collapses during a track are not handled. This
requires limit analysis of the polynomial fraction. It's perfectly
possible, but it is not implemented yet.