Added trackFoam test utility. A new particle class is also included to provide a simple wrapping around the base particle. At the moment it wraps the standard OpenFOAM particle. The intention is to write a new barycentric particle to wrap around instead.
Added findCell utility. This just reports the cell number of a given
location. Useful when writing lagragian positions files manually. This
might be unnecessary; there might be a utility that does this that I'm
not aware of. It'll get deleted eventually anyway.

Added a test case which is just a simple cube with a single track
through it.